backend/requirements.txt
fastapi==0.95.2
uvicorn==0.27.1
python-dotenv==1.0.0
supabase==0.7.1
google-generativeai==0.3.2
google-cloud-aiplatform==1.42.1
python-multipart==0.0.9
numpy==1.26.4
pydantic==1.10.12
faiss-cpu==1.7.4 